requirements.txt
Adafruit-Blinka==5.9.1
adafruit-circuitpython-dht==3.5.3
Adafruit-DHT==1.4.0
Adafruit-PlatformDetect==2.23.0
Adafruit-PureIO==1.1.8
asn1crypto==0.24.0
board==1.0
certifi==2018.8.24
chardet==3.0.4
colorzero==1.1
cryptography==2.6.1
entrypoints==0.3
flake8==3.6.0
future==0.18.2
gpiozero==1.5.1
idna==2.6
iso8601==0.1.13
keyring==17.1.1
keyrings.alt==3.1.1
mccabe==0.6.1
paho-mqtt==1.5.1
prometheus-client==0.9.0
pycodestyle==2.4.0
pycrypto==2.6.1
pyflakes==2.0.0
pyftdi==0.52.0
PyGObject==3.30.4
pyserial==3.5
python-apt==1.8.4.2
pyusb==1.1.0
pyxdg==0.25
PyYAML==5.3.1
requests==2.21.0
rpi-ws281x==4.2.5
RPi.GPIO==0.7.0
SecretStorage==2.3.1
serial==0.0.97
six==1.12.0
spidev==3.4
ssh-import-id==5.7
sysv-ipc==1.0.1
urllib3==1.24.1
